+++ lib/boring_backup/cli.rb
@@ -1,13 +1,13 @@
 require "thor"
 
-module NoopBackup
+module BoringBackup
   class CLI < Thor
     def self.exit_on_failure? = true
 
     desc "backup", "Create and store a new backup"
     def backup
-      NoopBackup.prepare!
-      result = NoopBackup::Commands::Backup.execute
+      BoringBackup.prepare!
+      result = BoringBackup::Commands::Backup.execute
       exit 1 unless result.success?
     rescue => e
       warn e