+++ lib/boring_backup/notifiers/stdout.rb
@@ -1,4 +1,4 @@
-module NoopBackup::Notifiers
+module BoringBackup::Notifiers
   class Stdout
     def notify(result)
       result.messages.each { |message| puts message }