+++ lib/boring_backup/stores/store.rb
@@ -1,0 +1,9 @@
+module BoringBackup::Stores
+  class Store
+    private
+
+    def config
+      @config ||= BoringBackup.config
+    end
+  end
+end