+++ lib/noop_backup/stores/store.rb
@@ -1,9 +1,0 @@
-module NoopBackup::Stores
-  class Store
-    private
-
-    def config
-      @config ||= NoopBackup.config
-    end
-  end
-end