+++ test/test_helper.rb
@@ -1,6 +1,6 @@
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
-require "noop_backup"
+require "boring_backup"
 require_relative "support/fake_store"
 
 require "minitest/autorun"